# Faster `square_clustering`