--- a/networkx/algorithms/bipartite/cluster.py
+++ b/networkx/algorithms/bipartite/cluster.py
@@ -261,9 +261,15 @@ def robins_alexander_clustering(G):
 
 def _four_cycles(G):
     cycles = 0
+    seen = set()
     for v in G:
-        for u, w in itertools.combinations(G[v], 2):
-            cycles += len((set(G[u]) & set(G[w])) - {v})
+        seen.add(v)
+        v_neighbors = set(G.neighbors(v))
+        two_hop_neighbors = set().union(*(G.neighbors(u) for u in v_neighbors))
+        two_hop_neighbors -= seen
+        for x in two_hop_neighbors:
+            p2 = len(v_neighbors & set(G.neighbors(x)))
+            cycles += p2 * (p2 - 1)
     return cycles / 4