--- a/networkx/algorithms/bipartite/cluster.py
+++ b/networkx/algorithms/bipartite/cluster.py
@@ -265,14 +265,14 @@ def _four_cycles(G):
     seen = set()
     for v in G:
         seen.add(v)
-        v_neighbors = set(G.neighbors(v))
+        v_neighbors = set(G._adj[v])
         if len(v_neighbors) < 2:
             # Can't form a square without at least two neighbors
             continue
-        two_hop_neighbors = set().union(*(G.neighbors(u) for u in v_neighbors))
+        two_hop_neighbors = set().union(*(G._adj[u] for u in v_neighbors))
         two_hop_neighbors -= seen
         for x in two_hop_neighbors:
-            p2 = len(v_neighbors.intersection(G.neighbors(x)))
+            p2 = len(v_neighbors.intersection(G._adj[x]))
             cycles += p2 * (p2 - 1)
     return cycles / 4