This implementation is based on my own toiling and experimentation. I don't know if this formulation is published anywhere. I named and described things as best I could, but I suspect it's still somewhat magical. I "discovered" this mostly from brute force and dumb luck--or a little intuition and a lot of perspiration--but it's still a bit mysterious to me. I believe it's solid though! The number of squares and triangles is based on https://arxiv.org/abs/2007.11111 ; the other terms in the denominator are more tricky.

Performance comparison (`*_prev` is original version, `*_new` is new version in this PR):
```python
In [1]: import networkx as nx

In [2]: G = nx.gnp_random_graph(5000, 0.001)
In [3]: %timeit nx.algorithms.cluster.square_clustering_prev(G)
307 ms ± 36.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [4]: %timeit nx.algorithms.cluster.square_clustering_new(G)
217 ms ± 31.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [5]: nx.algorithms.cluster.square_clustering_prev(G) == nx.algorithms.cluster.square_clustering_new(G)
Out[5]: True

In [6]: G = nx.complete_graph(100)

In [7]: %timeit nx.algorithms.cluster.square_clustering_prev(G)
5.54 s ± 197 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [8]: %timeit nx.algorithms.cluster.square_clustering_new(G)
60.6 ms ± 4.69 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
You can see we get a 90x speedup for the second example with the "massive" hundred node graph. I'm curious how it performs on larger, real-world graphs, which previously took _days_ when I benchmarked and compared to `graphblas_algorithms` (see the chart here: https://github.com/python-graphblas/graphblas-algorithms).

I changed the algorithm somewhat (and added a note) by ignoring self-loops (xref #7261 and #7238), which seems like the reasonable thing to do and matches other clustering algorithms.

We could probably get a factor of 2 or so performance if we cached `set(G.neighbors(v))` and perhaps more if we tried to reuse computation between iterations, but this would use a lot more memory.

CC @hagberg who first implemented `square_clustering` for NetworkX and @marcusjcrook who corrected it a few years ago (just in case y'all are interested in the fun of algorithm design!).