@eriknw 2025-01-26
How does NetworkX usually handle novel implementations? I know there's value in having good reference implementations that can be understood when reading source papers. I think the new version is a good reference implementation, but it is (probably) lacking a source paper (I haven't found one yet). Fwiw, I think this algorithm for square clustering is at least worth a conference paper; would anybody be interested in helping with this? If anybody finds relevant prior art--papers or implementations--please share!

Also, why is there not a `squares` function similar to `triangles`? It comes up: https://stackoverflow.com/questions/68893635/counting-the-number-of-squares-in-a-bipartite-networkx-graph