@eriknw 2025-01-27
Thanks for taking a look @rossbar!

The idea behind the new algorithm and why it's faster is straightforward: instead of iterating over all u-w combinations from v in search of u-v-w-x squares (and effectively visiting `x` many times), the new algorithm iterates over `x` (the two-hop neighbors) to find squares. From here, it's "just" a matter of counting and keeping the contributions straight (easier said than done!). Fwiw, I still find the original algorithm somewhat mysterious too.

While we're here, should we add `@not_implemented_for("directed")` to `square_clustering`? I don't think the original or new algorithms were ever meant to work on directed graphs, since there are no tests, special handling for directed, or discussion about them in the docstring as is done in `clustering`.