@eriknw 2025-01-30
Closely related, counting squares in bipartite graphs is often called "counting butterflies" or 2x2 bi-cliques in the literature. For example: ["Accelerated butterfly counting with vertex priority on bipartite graphs", Kai Wang, et al](https://d-nb.info/1263104584/34). For all kinds of graphs, 4-cycles is sometimes used.

It's pretty fun to tell people you were "counting butterflies" when they ask e.g. "what did you do last weekend?" 😜 

We may be able to use the same trick to speed up other places, such as:
(this first one can be sped up for sure)
https://github.com/networkx/networkx/blob/33768e175bad919cdf679329470581e078aeb444/networkx/algorithms/bipartite/cluster.py#L262-L267
https://github.com/networkx/networkx/blob/33768e175bad919cdf679329470581e078aeb444/networkx/algorithms/bipartite/redundancy.py#L109-L111
https://github.com/networkx/networkx/blob/33768e175bad919cdf679329470581e078aeb444/networkx/algorithms/centrality/dispersion.py#L63-L67
https://github.com/networkx/networkx/blob/33768e175bad919cdf679329470581e078aeb444/networkx/algorithms/community/quality.py#L312-L315