@eriknw 2025-02-11
I am collecting some benchmarks on real data here:

| data        | # nodes | # edges | # squares | prev  | new    | prev / new |
| -- | -- | -- | -- | -- | -- | -- |
| [email-Enron](https://sparse.tamu.edu/SNAP/email-Enron) | 36,692  | 367,662 | 145,048,916 | 254 s | 58.1 s | 4.4       |
| [preferentialAttachment](https://sparse.tamu.edu/DIMACS10/preferentialAttachment) | 100,000 | 999,970 | 360,332 | 78.3 s | 64.3 s | 1.2 |
| [caidaRouterLevel](https://sparse.tamu.edu/DIMACS10/caidaRouterLevel) | 192,244 | 1,218,132 | 192,960,428 | 152 s | 50.5 s | 3.0 |
| [dblp-2010](https://sparse.tamu.edu/LAW/dblp-2010) | 326,186 | 1,615,400 | 114,852,652 | 59 s | 17.3 s | 3.4 |
| [citationCiteseer](https://sparse.tamu.edu/DIMACS10/citationCiteseer) | 268,495 | 2,313,294 | 52,858,000 | 238 s | 145 s | 1.6 |
| [coAuthorsDBLP](https://sparse.tamu.edu/DIMACS10/coAuthorsDBLP) | 299,067 | 1,955,352 | 216,969,444 | 104 s | 38.3 s | 2.7 | 
| [as-Skitter](https://sparse.tamu.edu/SNAP/as-Skitter) | 1,696,415 | 22,190,596 | 251,076,792,072 | TODO | TODO | TODO |
| [coPapersCiteseer](https://sparse.tamu.edu/DIMACS10/coPapersCiteseer) | 434,102 | 32,071,440 | 943,803,419,816 | TODO | TODO | TODO |
| [coPapersDBLP](https://sparse.tamu.edu/DIMACS10/coPapersDBLP) | 5,404,486 | 30,491,458 | 197,618,846,172 | TODO | TODO | TODO |

Running the last three take _much_ longer than the others. My hunch is that the last three will also show the greatest improvement, because they have many more squares per node (or at least `coPapersCiteseer` will, because it has much higher average degree than the others and the most squares per node of all). Most of the data sets have average degree between 5-10, which is kind of boring that they're so similar. My hunch is based on my demonstration above that showed the new algorithm performing _much_ better on more dense graphs.

Heads up: the final benchmarks will take _days_ to run with the old implementation, so don't hold your breath. I computed number of squares for the last three with graphblas-algorithms.

If we optimize the previous implementation by using `G._adj`, we can get a factor of 2 or so improvement, which makes it faster than the new implementation in this PR for `preferentialAttachment` and `citationCiteseer` (barely) datasets.

Are there any other datasets (preferably from https://sparse.tamu.edu/) that anyone would like me to benchmark and compare?