@dschult 2025-04-03
Two comments from a quick read-through.
Why  is it better to iterate over the two-hop nbrs (x) of v (to find the combinations u-v) than to iterate over the combinations u-v to find the number of x? Are there fewer two-hop nbrs than combinations of u-v? Or is each lookup faster for finding u-v combinations given x than finding how many x exist for a u-v combination?   :)

This algorithm might be one where we could try making a dict-of-sets from the graph at the beginning -- throwing out selfloops and converting to sets so we don't have to do that repeatedly. We are doing the conversion already, so it won't it down for sure. How much it helps depends on how often we are repeating the lookup/conversion of neighbors to sets. But, in any case, we've done it elsewhere with success. Naming like `G_adj` makes the conversion from one to the other fairly easy -- but you'll want to remove the checking for selfloops code and maybe change other pieces too.  This is a kind of "since we're in the rabbit hole already" point to make. So its fine to say "enough for this PR".