@eriknw 2025-04-04
Thanks for taking a look!

> Why is it better to iterate over the two-hop nbrs (x) of v (to find the combinations u-v) than to iterate over the combinations u-v to find the number of x? Are there fewer two-hop nbrs than combinations of u-v? Or is each lookup faster for finding u-v combinations given x than finding how many x exist for a u-v combination? :)

Ha, great question, and one that is especially "fun" after being away from this for a couple months 😜. When I'm able to reply with a sensible explanation, I will... but not today. I think the benchmarks are pretty compelling though.

> This algorithm might be one where we could try making a dict-of-sets from the graph at the beginning -- throwing out selfloops and converting to sets so we don't have to do that repeatedly.

Yeah, in my original post of this thread I said "We could probably get a factor of 2 or so performance if we cached set(G.neighbors(v))", so getting a factor of say 50% is doable if we pre-compute a dict-of-sets. I didn't do this as part of this PR for two reasons: (1) doing so would use more memory, and (2) for `square_clustering` in particular, `nodes` can be given, so we may not need to convert all neighbors to sets. So, for now, I went for simply and less memory, but this could indeed be made a bit faster, and I wasn't sure if this tradeoff (more speed for more memory usage) was commonly done or acceptable for networkx.

> Naming like G_adj [...]

Heh, it turns out using `G_adj = G._adj` can save a non-trivial amount of time (5-10%)! Done in the latest commit.